Refactor
I've made a few changes to the library in order to support a pull-request I'll send soon that adds a feature to delete alerts. There were two approaches to refactor the code in order to avoid method name collision:

1.- Create a new `Client` class and move the request methods to it.
2.- Remove the request methods and just use the generic `#request`.

I prefer the latter way because it's simpler and I think it's an elegant way to solve the problem.